+++ enclave/sort/quick.cpp
@@ -2,15 +2,15 @@
 std::random_device dev2;
 std::mt19937 rng2(dev2());
 
-int64_t partition(int64_t *arr, int64_t low, int64_t high) {
+int partition(int *arr, int low, int high) {
   // TODO: random version
   // srand(unsigned(time(NULL)));
-  std::uniform_int_distribution<int64_t> dist{low, high};
-  int64_t randNum = dist(rng2);
+  std::uniform_int_distribution<int> dist{low, high};
+  int randNum = dist(rng2);
   swapRow(arr + high, arr + randNum);
-  int64_t *pivot = arr + high;
-  int64_t i = low - 1;
-  for (int64_t j = low; j <= high - 1; ++j) {
+  int *pivot = arr + high;
+  int i = low - 1;
+  for (int j = low; j <= high - 1; ++j) {
     if (cmpHelper(pivot, arr + j)) {
       i++;
       if (i != j) {
@@ -24,14 +24,14 @@ int64_t partition(int64_t *arr, int64_t low, int64_t high) {
   return (i + 1);
 }
 
-int64_t partition(Bucket_x *arr, int64_t low, int64_t high) {
-  std::uniform_int_distribution<int64_t> dist{low, high};
-  int64_t randNum = dist(rng2);
+int partition(Bucket_x *arr, int low, int high) {
+  std::uniform_int_distribution<int> dist{low, high};
+  int randNum = dist(rng2);
   // int randNum = rand() % (high - low + 1) + low;
   swapRow(arr + high, arr + randNum);
   Bucket_x *pivot = arr + high;
-  int64_t i = low - 1;
-  for (int64_t j = low; j <= high - 1; ++j) {
+  int i = low - 1;
+  for (int j = low; j <= high - 1; ++j) {
     if (cmpHelper(pivot, arr + j)) {
       i++;
       if (i != j) {
@@ -46,18 +46,18 @@ int64_t partition(Bucket_x *arr, int64_t low, int64_t high) {
 }
 
 
-void quickSort(int64_t *arr, int64_t low, int64_t high) {
+void quickSort(int *arr, int low, int high) {
   if (high > low) {
-    int64_t mid = partition(arr, low, high);
+    int mid = partition(arr, low, high);
     quickSort(arr, low, mid - 1);
     quickSort(arr, mid + 1, high);
   }
 }
 
-void quickSort(Bucket_x *arr, int64_t low, int64_t high) {
+void quickSort(Bucket_x *arr, int low, int high) {
   if (high > low) {
     // printf("In quickSort: %lu, %lu\n", low, high);
-    int64_t mid = partition(arr, low, high);
+    int mid = partition(arr, low, high);
     quickSort(arr, low, mid - 1);
     quickSort(arr, mid + 1, high);
   }
